Add output::OutputSender.

+++ src/command_line_args.rs
@@ -6,9 +6,7 @@ use tokio::sync::OnceCell;
 
 use tracing::debug;
 
-fn default_jobs() -> u32 {
-    // Using u32 instead of usize for jobs argument due to this issue:
-    // https://github.com/tokio-rs/tokio/issues/4446#issuecomment-1365930467
+fn default_jobs() -> u64 {
     num_cpus::get().try_into().unwrap()
 }
 
@@ -26,8 +24,8 @@ pub struct CommandLineArgs {
     pub input: Vec<String>,
 
     /// Maximum number of commands to run in parallel, defauts to num cpus
-    #[arg(short, long, default_value_t = default_jobs(), value_parser = clap::value_parser!(u32).range(1..))]
-    pub jobs: u32,
+    #[arg(short, long, default_value_t = default_jobs(), value_parser = clap::value_parser!(u64).range(1..))]
+    pub jobs: u64,
 
     /// Use null separator for reading input instead of newline.
     #[arg(short('0'), long)]
